backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pandas==2.1.3
scikit-learn==1.3.2
lazypredict==0.2.12
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
alembic==1.13.0
python-dotenv==1.0.0
openai==1.3.7
plotly==5.17.0
pymupdf==1.23.8
numpy==1.24.3
matplotlib==3.8.2
seaborn==0.13.0
python-dateutil==2.8.2
aiofiles==23.2.1
httpx==0.25.2 